Update helper and data for context
